+++ cryptnoxpy/card/__init__.py
@@ -5,3 +5,4 @@ Module that contains classes for various cryptnox cards.
 from .base import Base
 from .basic_g0 import BasicG0
 from .basic_g1 import BasicG1
+from .nft import Nft