+++ cryptnoxpy/exceptions.py
@@ -40,6 +40,9 @@ class DerivationSelectionException(CryptnoxException):
     """Not a valid derivation selection."""
 
 
+class KeySelectionException(CryptnoxException):
+    """Not a valid key type selection"""
+
 class EOSKeyError(CryptnoxException):
     """The signature wasn't compatible with EOS standard after 10 tries"""